refactoring

+++ printf/ft_convert_to_base.c
@@ -6,13 +6,13 @@
 /*   By: kvalerii <kvalerii@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/20 10:42:32 by kvalerii          #+#    #+#             */
-/*   Updated: 2024/10/22 11:27:32 by kvalerii         ###   ########.fr       */
+/*   Updated: 2024/10/30 11:31:44 by kvalerii         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-int	ft_convert_to_base(unsigned long dec, char *base, int base_len)
+int	ft_convert_to_base(unsigned long dec, char *base, size_t base_len)
 {
 	char	c;
 	int		a;
