+++ printf/ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: kvalerii <kvalerii@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/18 21:04:10 by valeriia          #+#    #+#             */
-/*   Updated: 2024/10/30 10:55:36 by kvalerii         ###   ########.fr       */
+/*   Updated: 2024/10/30 11:31:59 by kvalerii         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,7 @@
 # define FT_PRINTF_H
 
 # include "./libft/libft.h" 
-# include "stdarg.h"
-# include <limits.h>
+# include <stdarg.h>
 
 int	ft_printf(const char *format, ...);
 int	ft_printf_c(char a);
@@ -27,6 +26,6 @@ int	ft_printf_u(unsigned int dec);
 int	ft_printf_x(unsigned long dec);
 int	ft_printf_ux(unsigned long dec);
 int	ft_printf_p(unsigned long dec);
-int	ft_convert_to_base(unsigned long dec, char *base, int base_len);
+int	ft_convert_to_base(unsigned long dec, char *base, size_t base_len);
 
 #endif
